fix(opentracing-shim): register tracer provider for complete traces
Traces generated by the opentracing shim example were incomplete. One trace was generated per service instead
of a single trace containing all spans from both services for the remote calls. The registration of the provider appeared
to have been done in the other opentelemetry examples but not this one. After adding the same change a single trace
is generated as expected.