fix: ignore TIMERWRAP in AsyncHooksContextManager
TIMERWRAP is used internally in node.js up to version 11 to combine
timeouts with the same timeout value. This combination can lead to
false context propagation.

Every timer additionally creates a Timeout resource therefore TIMERWRAP
is not needed for correct propagation and it's safe to ignore it in
context manager.