fix: do not propagate parent spand id from b3
B3 has an option to propagate both span id and a parent span id so that
both sides of request can share the same span id. This is a usecase that
we do not support in OpenTelemetry and couldn't support even if we wanted
to because parent span id is not readable outside of the export
pipeline. This buggy implementation stores and propagates the extracted
parent span id, which will break zipkin implementations as a result.