chore: requires user to pass context to propagation APIs
At least for extract() using the current active context is often a bad
choice because usually a plugin wants to use an extracted span instead
the current active.

Change propagation APIs to require context as first argument instead of
using active context on default. This ensures that plugin developers
have to be explicit instead of relying on a potential wrong default.