chore: improve naming of span related context APIs
Rename [gs]etActiveSpan() to [gs]etSpan() as the span is not active anywhere,
it's just referenced by the context.

Rename setExtractedSpanContext() to setSpanContext() because any SpanContext
may be set not only extracted ones.

Rename getParentSpanContext() to getSpanContext() because it retrieves the
context of the span which can be any span not only a parent.