fix: ensure we generate against v1.1.0 of the spec
Changed `generate.sh` to ensure it generates the files against the latest
published version of the otel specification which is v1.1.0.

To allow this version of the spec files to work with the `semconv` the
version of this tool had to be lowered to 0.2.1 as there is a breaking
change in 0.3.0 which removes support for the `number`-type.