chore: cleanup setting config in instrumentations
There is no need to create a copy of the config object as this is done in base class.
Remove defaulting to an empty config object at several places as this is done in the base class.
Remove the intersection type with InstrumentationConfig as all config types extend this type.
Avoid overriding the _config property in the grpc instrumentation.