refactor(instrumentation-http): set span status once on response errors
On the error listener, `utils.setSpanWithError()` was called, which
already sets the span status as ERROR, and then, for some special cases,
the response code was parsed and set the result code on the same span,
modifying the same span status twice.

`utils.setSpanWithError()` now accepts an optional parameter with the
status code in case the default one (Error) is not wanted.