Generate named exports for semantic attributes
SemanticAttributes and SemanticResourceAttributes export an object with
a large number of verbose keys and values which is not tree-shakeable,
so when a consumer imports one key they import them all.

When used in a browser context, this can be quite costly because the
entire SemanticAttributes/SemanticResourceAttributes object must be sent
over the wire.

By exporting named constants for the individual keys and values as well,
we can support tree shaking to reduce this size.