refactor meter/tracer provider creation
this moves the creation and registration of each into their own
function and reduces some of the existence checks on the provider
configurations. the built-in opinionated tracer/meter provider
will only be created if it was explicitly configured by the sdk
consumer. otherwise there is default behavior:

1. for meters, the default behavior is to have no meterprovider.
2. for tracers, the default behavior is to configure a tracer
   provider, potentially with multiple exporters, from env.

this commit does not change the above behavior, this message is
just documenting how it works.