feat(sdk-node): tracer provider factory via config
currently, if the sdk consumer chooses to configure the tracer-specific
config options, an opinionated tracer will be created for them using the
batch processor or exporter they specified.

if no such config is provided the default behavior is to use the
TracerProviderWithEnvExporter, which provides another opinionated implementation
of possibly multiple export formats derived from env.

however, sdk consumers may desire to configure the tracer provider themselves,
e.g., to export to two different otlp backends but utilizing that protocol
for each---this cannot currently be done.

this commit adds support for specifying a factory method, which will be
invoked on start()---the factory method receives the Resource that was
detected in the start function, and creates a tracer provider from that input.
sdk consumers should be able to utilize this factory function to set any
additional configuration they desire.