use --experimental-meta-resolve option with hook file (#8)
Allows `hook.mjs` file to exported from the instrumentation package if
the user also uses the `--experimental-meta-resolve` flag. If a user
doesn't want to use this flag, they can still reference
`import-in-the-middle/hook.mjs`. I figure this option is safe because
this is such a Node targeted feature that I'm not worried about someone
trying to run this in the browser. The code path to the ESM hook is
entirely contained in the `node` folder which wouldn't be available if
someone were to run this package in the browser and they wouldn't be
using the hook anyway.

---------

Co-authored-by: Jamie Danielson <jamieedanielson@gmail.com>