tests: regenerate recently-expired test TLS certs (#4777)
* tests: regenerate recently-expired test TLS certs

Tests started failing yesterday because the various TLS certs in
the repo used for TLS testing had expired

$ openssl x509 -noout -enddate -in experimental/packages/exporter-logs-otlp-grpc/test/certs/ca.crt
notAfter=Jun  7 16:44:23 2024 GMT

This change regenerates all the TLS certs used for testing and
examples/https. It also adds a test case in each relevant package
that will fail with an obvious message when these certs expire
next time (certs are generated with a 1y expiry).

This also adds a comment 'npm run main:regenerate-test-certs'
that can be used to run the maintenance task.

* lint:fix, add changelog entry

* skip cert expiry test on node v14, because crypto.X509Certificate came in node v15