add dummy entry to release-notes script for new pkg
The regex checks for existence of the previous h2
but new packages will not have versions before it.
Add a dummy h2 to satisfy the regex with minimal
overall changes.
To test this, replace Unreleased in the CHANGELOG
to an actual version, like 1.1.1.
Then run the script as described in the comments.
There should be a .tmp/release-notes.md with
the contents of the changelog for that version.