[sdk-metrics] Switch workaround types to official @otel/api types
These types were introduced in the 1.0 series when the metrics
package depended on an older version of the api package that didn't
come with those types, so the interfaces were essentially vendored
into the package for internal use.

Now with the 2.0 release on the horrizon, we can bump the required
version of @otel/api and is now safe to unvendor these types in
favor of the official ones.

Closes #5223
Closes #5224