refactor(sdk-metrics) Swap workaround types for @otel/api types
The internal `Gauge` and the unnamed `MetricAdvice` types were
introduced in the 1.0 series when the metrics package depended on
an older version of the api package that didn't come with those
types, so the interfaces were essentially vendored into the package
for internal use.

Now with the 2.0 release on the horrizon, we can bump the required
version of @otel/api and is now safe to unvendor these types in
favor of the official ones.

This is not expected to have any impact on consumers as the types
are intended to be compatible for the positions they were used in.

However, the minimum requirement for the `@opentelemetry/api` peer
dependency has been raised to 1.9.0.

Fixes #5223
Fixes #5224