Port the first set of actual fetch() test to new msw test infra
Tightened some of the assertions in the old tests where they were
unnecessarily vague (e.g. protocol is either "http" or "https"
when the test setup already fully dictated which one will be used).

This shows the promise of the new appraoch – we are able to test
the instrumentation without mocking the browser APIs at all, fully
execrising the instrumentation code including `PerformanceObserver`
and the resource-timing matching code.

We did cheat and added a 500ms delay to make this work, so we do
have to come back and deal with that at some later point.