Port CORS tests
As pointed out #5122, the expectations in the existing tests are
not the real browser behavior (which was ultimately the motivation
for refactoring these tests to rely on real browser behavior as
much as possible).

So instead of porting them as-is, this adjust the tests behavior to
match what actually happens.