refactor(api)!: Use TypeScript overload signature for `addEvent`
This makes the intent more clear, and makes the invalid signature
`addEvent(name, timeStamp1, timeStamp2)` explicitly illegal in TS.

This has always been the intent, and the textual documentation
somewhat implies that already.

Technically, this is breaking change for TypeScript users. There
are situations previously valid TS code will stop being accepted
by the type checker, even if the code is semantically valid. It
mostly arises when the caller is trying to propagate the same-ish
overload signature. It can still be done safely, just requires a
bit more care.

In any case, this will go into the 2.0 release so it's a great
time to make this change.