feat(api)!: Revamp `TextMapPropagator` TypeScript interface
The previous interface placed the generic paramter on the interface
itself. This implies that the implementors of `TextMapPropagator`
can specify what carrier types it accepts.

As far as I can tell, this has never been the case.

The propagator API is designed to be called from the transport
layers. The transport layer (caller of `inject`/`extract`) is the
one who has control over what the carrier type is. The constraint
is that the carrier must semantically behave as an abstract map
data structure that supports setting and getting string keys/values,
and the caller must supply a matching getter/setter that works with
the given carrier type.

For example, the fetch instrumentation calls the propogation API
with one of many carrier types – POJOs, `Headers`, `Map`, etc.

Therefore, a _correct_ implementation of `TextMapPropagator` must
treat the carrier as opaque and only work with it using the supplied
getter/setter.

Unfortunately, the previous interface definition allowed a lot of
sloppiness in the implementations that would cause problems in the
real world. Fortunately, it seems like these all happen in tests,
and all the production propagators are already compliant with the
spirit of the API.

This commit moves the generic parameter from the interface into
the `inject` and `extract` methods. This forces implementors to
comply with the API more rigiously.

This is a breaking change even for compliant implementations, as
it requires adjusting the method signatures to match. If an
implementation supplied a custom type for the generic parameter
previously existed on the interface, then that probably represents
a case where it wouldn't have worked in the real world and should
be refactored/rewritten to work with the abstracted carrier type
as intended.

Ref #5365