refactor(sdk-trace-node): fix eslint warning
```
/home/runner/work/opentelemetry-js/opentelemetry-js/packages/opentelemetry-sdk-trace-node/test/registration.test.ts
  30:61  warning  Don't use `Function` as a type  @typescript-eslint/ban-types
```

Here we are looking for a `AnyConstructor` type, and `Function` is
a close enough approximation that exists in the standard library.

The potential foot gun of _calling_ the function resulting in an
`any` value still exists, but this is a small function used in
tests only, so that's probably acceptable.

Ref #5365