refactor(sdk-trace-web): fix eslint warning
```
/home/runner/work/opentelemetry-js/opentelemetry-js/packages/opentelemetry-sdk-trace-web/src/utils.ts
  52:14  warning  Unexpected any. Specify a different type  @typescript-eslint/no-explicit-any
```

The eslint comment wasn't working because it was placed on the
wrong line. However, it also wasn't necessary anymore – `keyof any`
was used to obtain the union of all possible object key types –
`string | number | symbol`. TypeScript now provides an alias for
this in the standard libaray that gives exactly the same result.

Also fixed what appears to be a TypeError on an un-annotated `let`
binding, not sure why it wasn't caught by the type checker, maybe
it was just missed on the version of TS we are using.

Ref #5365