fix(exporter-zipkin): remove deprecated `url.parse` usage
Addresses the following deprecation/eslint warning:

```
/home/runner/work/opentelemetry-js/opentelemetry-js/packages/opentelemetry-exporter-zipkin/src/platform/node/util.ts
  34:19  warning  'url.parse' was deprecated since v11.0.0. Use 'url.URL' constructor instead  node/no-deprecated-api
```

`http(s).request` now takes a URL (the global one, not `url.URL`)
argument separate from the options bag, so we should be able to
make this change compatibly, outside of some weird edge cases that
are probably not applicable here.

It is worth noting in the CHANGELOG since this used to trigger a
Node runtime deprecation warning with `--pending-deprecation`.

Ref #5365