refactor(exporter-metrics-otlp-http): fix eslint warning
Fixes the following eslint warning in exporter-metrics-otlp-http:

```
/home/runner/work/opentelemetry-js/opentelemetry-js/experimental/packages/opentelemetry-exporter-metrics-otlp-http/src/OTLPMetricExporterBase.ts
  120:30  warning  Unexpected any. Specify a different type  @typescript-eslint/no-explicit-any
```

The unused parameter can be explicitly typed as `InstrumentType`,
but since this default selector doesn't actually care about it, it
can also just be elided.

By adding an explicit return type, if someone in the future added
the wrong parameter to the default selector function, TypeScript
would reject it.

Ref #5365