refactor(exporter-trace-otlp-http): fix eslint warnings
```
/home/runner/work/opentelemetry-js/opentelemetry-js/experimental/packages/exporter-trace-otlp-http/src/index.ts
  20:1  warning  Using 'ExportAllDeclaration' is not allowed  no-restricted-syntax

/home/runner/work/opentelemetry-js/opentelemetry-js/experimental/packages/exporter-trace-otlp-http/src/platform/browser/index.ts
  20:1  warning  Using 'ExportAllDeclaration' is not allowed  no-restricted-syntax

/home/runner/work/opentelemetry-js/opentelemetry-js/experimental/packages/exporter-trace-otlp-http/src/platform/index.ts
  20:1  warning  Using 'ExportAllDeclaration' is not allowed  no-restricted-syntax

/home/runner/work/opentelemetry-js/opentelemetry-js/experimental/packages/exporter-trace-otlp-http/src/platform/node/index.ts
  20:1  warning  Using 'ExportAllDeclaration' is not allowed  no-restricted-syntax
```

The existing eslint comment wasn't working because it was in the
wrong format.

The comment say we should do this before the next "major" (minor
since we are pre-1.0?) version, which means this is the right time
to do it.

However, I'm not sure if there is a real need/reason for that. As
far as I can tell, this is still exporting exactly the same items
and shouldn't cause any breakages?