refactor(exporter-metrics-otlp-proto): fix eslint warning
```
/home/runner/work/opentelemetry-js/opentelemetry-js/experimental/packages/opentelemetry-exporter-metrics-otlp-proto/src/index.ts
  20:1  warning  Using 'ExportAllDeclaration' is not allowed  no-restricted-syntax
```

The comment say we should do this before the next "major" (minor
since we are pre-1.0?) version, which means this is the right time
to do it.

However, I'm not sure if there is a real need/reason for that. As
far as I can tell, this is still exporting exactly the same items
and shouldn't cause any breakages?

Ref #5365