chore: add workflow for surveys of non-members
As discusses on a SIG meeting a few weeks back, the Contributor Experience is adding
surveys to PRs merged by non-members, to gather info about their experiences. We're
starting to add to a couple of repos for testing. This PR adds to this repo.

An example of this message can be found here: https://github.com/open-telemetry/opentelemetry.io/pull/6099#issuecomment-2619960339