chore: unpin semantic-conventions package dep
There is no need to pin deps for the semantic-conventions package.
Our recommendation is to not pin, to allow for de-duplication of installs
of the package in node_modules trees to reduce install size.
The concerns that motivate pinning other otel deps, described in
https://github.com/open-telemetry/opentelemetry-js/issues/5283, do
not apply to the semantic-conventions package.

This also removes the updating of the semconv deps when the semconv
package is released. There should be no need to do so. A particular
package can bump its minimum semconv dep if/when it uses exports
in a newer version.