feat(instrumentation-http): Add back support for http semconv
At this point it is mostly copy-paste from when
most of it was removed, but there have been a
few features since then (synthetic) and also a
few things were dropped and re-added already
like metrics. This also uses the new utility for
semconv stability instead of the old hardcoded enum.
Some things are still missing according to tests so will
need another pass. Doing this all in one PR is... a lot.
Possibly too much, but not sure if it is easier to
split either. It is just a lot of change.